slave_server/requirements.txt
django<2
pyyaml
redis
django-cors-headers
pyserial
requests
